test: split the three slowest files so the Windows shards can balance
The 3-way shard run (32863905788) landed at 9m35s / 18m56s / 13m58s
against a <=15 minute target because vitest's `--shard` slices by FILE
and three files were indivisible blocks on the Windows leg:
flowLegKnownBad.test.ts (661s), scenarios.test.ts (422s), and
webviewFixtureGate.test.ts (294s). Whatever shard drew the first of
those carried an ~11-minute floor no shard count could cut.

Each becomes a set of smaller files behind a shared non-test module,
so the slices can spread:

- flowLegKnownBad: one file per known-bad leg, assertions
  single-sourced in flowLegKnownBadHarness.ts.
  flowLegKnownBad.coverage.test.ts pins one leg file per leg id,
  restoring the enumeration guarantee the old single-file loop
  provided structurally.
- scenarios: local-states / remote-and-shelf / independence, split by
  prepare() cost; the independent git read-back seam and the scratch
  registries move to scenariosTestHelpers.ts unchanged.
- webviewFixtureGate: core / update / disposal; shared fixtures move
  to webviewFixtureGateTestHelpers.ts. Each file keeps its own
  vi.mock("vscode") block because vitest hoists mocks per test file.

Every test body, title, and timeout is carried over verbatim: the old
files held 4 + 57 + 23 tests and the split files run 84 + 1 new
coverage test, all green locally.
