+++ tests/unit/e2e/flowLegKnownBad.durable-state.test.ts
@@ -1,0 +1,5 @@
+// One leg per file so vitest's file-count `--shard` can spread the four subprocess runs across
+// the Windows CI shards -- see flowLegKnownBadHarness.ts for the full story and the shared suite.
+import { flowLegKnownBadSuite } from "./flowLegKnownBadHarness";
+
+flowLegKnownBadSuite("durable-state");
